Consolidated goto support machinery in hir-analysis
