+++ crates/language-server/test_files/goto.snap.new
@@ -1,0 +1,33 @@
+---
+source: crates/language-server/src/functionality/goto.rs
+assertion_line: 258
+expression: snapshot
+input_file: test_files/goto.fe
+---
+0: use core
+1: 
+2: struct Foo {}
+3: struct Bar {}
+4: 
+5: fn main() {
+6:     let x: Foo
+7:     let y: Bar
+8:     let z: baz::Baz
+9:     core::todo()
+10: }
+11: 
+12: mod baz {
+13:    pub struct Baz {}
+14: }
+---
+cursor position (1, 5), path: lib
+cursor position (3, 8), path: goto::Foo
+cursor position (4, 8), path: goto::Bar
+cursor position (6, 4), path: goto::main
+cursor position (7, 12), path: goto::Foo
+cursor position (8, 12), path: goto::Bar
+cursor position (9, 12), path: goto::baz
+cursor position (9, 17), path: goto::baz::Baz
+cursor position (10, 5), path: lib
+cursor position (10, 11), path: lib::todo
+cursor position (14, 15), path: goto::baz::Baz
