+++ crates/language-server/test_files/goto_debug.snap.new
@@ -1,0 +1,22 @@
+---
+source: crates/language-server/src/functionality/goto.rs
+assertion_line: 258
+expression: snapshot
+input_file: test_files/goto_debug.fe
+---
+0: struct Point {
+1:     pub x: i32,
+2:     pub y: i32
+3: }
+4: 
+5: fn test() {
+6:     let p = Point { x: 1, y: 2 }
+7:     let val = p.x
+8: }
+---
+cursor position (1, 8), path: goto_debug::Point
+cursor position (2, 9), path: goto_debug::Point::x
+cursor position (3, 9), path: goto_debug::Point::y
+cursor position (6, 4), path: goto_debug::test
+cursor position (7, 13), path: goto_debug::Point
+cursor position (8, 17), path: goto_debug::Point::x
