+++ crates/language-server/test_files/goto_field_test.snap.new
@@ -1,0 +1,20 @@
+---
+source: crates/language-server/src/functionality/goto.rs
+assertion_line: 258
+expression: snapshot
+input_file: test_files/goto_field_test.fe
+---
+0: struct MyStruct {
+1:     pub field: u32
+2: }
+3: 
+4: fn main() {
+5:     let obj = MyStruct { field: 42 }
+6:     let val = obj.field
+7: }
+---
+cursor position (1, 8), path: goto_field_test::MyStruct
+cursor position (2, 9), path: goto_field_test::MyStruct::field
+cursor position (5, 4), path: goto_field_test::main
+cursor position (6, 15), path: goto_field_test::MyStruct
+cursor position (7, 19), path: goto_field_test::MyStruct::field
