+++ crates/language-server/test_files/goto_values.snap.new
@@ -1,0 +1,71 @@
+---
+source: crates/language-server/src/functionality/goto.rs
+assertion_line: 258
+expression: snapshot
+input_file: test_files/goto_values.fe
+---
+0: use core
+1: 
+2: struct MyStruct {
+3:     pub field: u32
+4: }
+5: 
+6: const MY_CONST: u32 = 42
+7: 
+8: fn my_function() -> u32 {
+9:     MY_CONST + 10
+10: }
+11: 
+12: fn another_function(param: MyStruct) -> u32 {
+13:     param.field + my_function()
+14: }
+15: 
+16: fn main() {
+17:     let x: MyStruct = MyStruct { field: 5 }
+18:     let y = my_function()
+19:     let z = another_function(x)
+20:     let c = MY_CONST
+21:     
+22:     core::todo()
+23: }
+24: 
+25: mod nested {
+26:     pub const NESTED_CONST: u32 = 100
+27:     
+28:     pub fn nested_function() -> u32 {
+29:         NESTED_CONST * 2
+30:     }
+31: }
+32: 
+33: fn test_nested() {
+34:     let a = nested::NESTED_CONST
+35:     let b = nested::nested_function()
+36: }
+---
+cursor position (1, 5), path: lib
+cursor position (3, 8), path: goto_values::MyStruct
+cursor position (4, 9), path: goto_values::MyStruct::field
+cursor position (7, 7), path: goto_values::MY_CONST
+cursor position (9, 4), path: goto_values::my_function
+cursor position (10, 5), path: goto_values::MY_CONST
+cursor position (13, 4), path: goto_values::another_function
+cursor position (13, 21), path: goto_values::another_function::param
+cursor position (13, 28), path: goto_values::MyStruct
+cursor position (14, 5), path: goto_values::another_function::param
+cursor position (14, 11), path: goto_values::MyStruct::field
+cursor position (14, 19), path: goto_values::my_function
+cursor position (17, 4), path: goto_values::main
+cursor position (18, 12), path: goto_values::MyStruct
+cursor position (18, 23), path: goto_values::MyStruct
+cursor position (19, 13), path: goto_values::my_function
+cursor position (20, 13), path: goto_values::another_function
+cursor position (21, 13), path: goto_values::MY_CONST
+cursor position (23, 5), path: lib
+cursor position (23, 11), path: lib::todo
+cursor position (27, 15), path: goto_values::nested::NESTED_CONST
+cursor position (29, 12), path: goto_values::nested::nested_function
+cursor position (30, 9), path: goto_values::nested::NESTED_CONST
+cursor position (34, 4), path: goto_values::test_nested
+cursor position (35, 13), path: goto_values::nested
+cursor position (36, 13), path: goto_values::nested
+cursor position (36, 21), path: goto_values::nested::nested_function
